renaming files with a clear name
